Move markdown from `hover` to `ide`
commit-id:f64daa1d
