+++ src/ide/hover/render/definition.rs
@@ -9,7 +9,7 @@ use cairo_lang_syntax::node::{TypedStablePtr, TypedSyntaxNode};
 use lsp_types::Hover;
 
 use crate::ide::hover::markdown_contents;
-use crate::ide::hover::render::markdown::{RULE, fenced_code_block};
+use crate::ide::markdown::{RULE, fenced_code_block};
 use crate::lang::db::AnalysisDatabase;
 use crate::lang::defs::{ResolvedItem, SymbolDef};
 use crate::lang::lsp::ToLsp;
