+++ src/ide/hover/render/mod.rs
@@ -4,5 +4,4 @@ pub use self::ty::*;
 
 mod definition;
 mod literal;
-mod markdown;
 mod ty;
